Implement q_ascend
To use the built-in macro for safely looping through each entry in
queue, the order of queue need to be reversed first. Whenever
encountering value larger than the last entry in queue (prior-reversed),
delete the node from queue and free the resources related. If the
encountered node has a smaller value, update the minimum and increment
counts of the queue size.