Fix wrong counting of queue size
Last time, nodes with the same value as minimum would NOT be counted.
The seperation of condition "equal" and condition "less than" can be
merged into one combined condition "less than or equal to" that can be ommited
since if one value is not greater the other, it must be "less than or
equal to". When bumpping into entris with equal value, there shuold be
no harm to replace the original minimum with the new one.