Implement q_descend
Check if queue is not empty and valid before further operation on it.
First reverse the queue so the macro list_for_each_entry_safe can be
used to safely travese through the queue since the last entry in the
queue (prior-reversed) is considered a pivot implying a backward
walk-trouth. Any entry encountered with a smaller value gets eliminated
since the value for the pivot should be the smallest. Count up the size
of queue by one each time an entry with larger value is hit and update
the maximum with the new one.